move DroneOptions
