+++ src/unofficial_wg_implementations.rs
@@ -5,7 +5,7 @@
 use std::collections::HashMap;
 
 use crossbeam_channel::{Receiver, Sender};
-use wg_2024::{config::Config, controller::Command, network::NodeId};
+use wg_2024::{config::Config, controller::Command, network::NodeId, packet::Packet};
 
 pub struct SimControllerOptions {
     pub command_send: HashMap<NodeId, Sender<Command>>,
@@ -16,4 +16,16 @@ pub struct SimControllerOptions {
 pub trait SimulationController{
     fn new(opt:SimControllerOptions)-> Self;
     fn run(&mut self);
+}
+
+// currently is in working group PR
+#[derive(Debug, Clone)]
+
+pub struct DroneOptions {
+    pub id: NodeId,
+    pub sim_contr_send: Sender<Command>,
+    pub sim_contr_recv: Receiver<Command>,
+    pub packet_recv: Receiver<Packet>,
+    pub packet_send: HashMap<NodeId, Sender<Packet>>,
+    pub pdr: f32,
 }
